fix error in parsing timeline
