--- downloading/user.go
@@ -1,105 +1,0 @@
-package downloading
-
-import (
-	"context"
-
-	"github.com/go-resty/resty/v2"
-	"github.com/jmoiron/sqlx"
-	log "github.com/sirupsen/logrus"
-	"github.com/unkmonster/tmd2/database"
-	"github.com/unkmonster/tmd2/internal/utils"
-	"github.com/unkmonster/tmd2/twitter"
-)
-
-// 更新数据库中对用户的记录
-func syncUser(db *sqlx.DB, user *twitter.User) error {
-	renamed := false
-	isNew := false
-	usrdb, err := database.GetUserById(db, user.Id)
-	if err != nil {
-		return err
-	}
-
-	if usrdb == nil {
-		isNew = true
-		usrdb = &database.User{}
-		usrdb.Id = user.Id
-	} else {
-		renamed = usrdb.Name != user.Name || usrdb.ScreenName != user.ScreenName
-	}
-
-	usrdb.FriendsCount = user.FriendsCount
-	usrdb.IsProtected = user.IsProtected
-	usrdb.Name = user.Name
-	usrdb.ScreenName = user.ScreenName
-
-	if isNew {
-		err = database.CreateUser(db, usrdb)
-	} else {
-		err = database.UpdateUser(db, usrdb)
-	}
-	if err != nil {
-		return err
-	}
-	if renamed || isNew {
-		err = database.RecordUserPreviousName(db, user.Id, user.Name, user.ScreenName)
-	}
-	return err
-}
-
-func getTweetAndUpdateLatestReleaseTime(ctx context.Context, client *resty.Client, user *twitter.User, entity *UserEntity) ([]*twitter.Tweet, error) {
-	tweets, err := user.GetMeidas(ctx, client, &utils.TimeRange{Min: entity.LatestReleaseTime()})
-	if err != nil || len(tweets) == 0 {
-		return nil, err
-	}
-	if err := entity.SetLatestReleaseTime(tweets[0].CreatedAt); err != nil {
-		return nil, err
-	}
-	return tweets, nil
-}
-
-func DownloadUser(ctx context.Context, db *sqlx.DB, client *resty.Client, user *twitter.User, dir string) ([]PackgedTweet, error) {
-	if user.Blocking || user.Muting {
-		return nil, nil
-	}
-
-	_, loaded := syncedUsers.Load(user.Id)
-	if loaded {
-		log.WithField("user", user.Title()).Debugln("skiped downloaded user")
-		return nil, nil
-	}
-	entity, err := syncUserAndEntity(db, user, dir)
-	if err != nil {
-		return nil, err
-	}
-
-	syncedUsers.Store(user.Id, entity)
-	tweets, err := getTweetAndUpdateLatestReleaseTime(ctx, client, user, entity)
-	if err != nil || len(tweets) == 0 {
-		return nil, err
-	}
-
-	// 打包推文
-	pts := make([]PackgedTweet, 0, len(tweets))
-	for _, tw := range tweets {
-		pts = append(pts, TweetInEntity{Tweet: tw, Entity: entity})
-	}
-
-	return BatchDownloadTweet(ctx, client, pts...), nil
-}
-
-func syncUserAndEntity(db *sqlx.DB, user *twitter.User, dir string) (*UserEntity, error) {
-	if err := syncUser(db, user); err != nil {
-		return nil, err
-	}
-	expectedTitle := utils.WinFileName(user.Title())
-
-	entity, err := NewUserEntity(db, user.Id, dir)
-	if err != nil {
-		return nil, err
-	}
-	if err = syncPath(entity, expectedTitle); err != nil {
-		return nil, err
-	}
-	return entity, nil
-}
